+++ src/utils/setTerminalTitle.ts
@@ -1,5 +1,0 @@
-export default function setTerminalTitle(title: string) {
-	process.stdout.write(
-		`${String.fromCharCode(27)}]0;${title}${String.fromCharCode(7)}`,
-	);
-}